Create domain job handler, set default enum conc limit
Create and Update domains are scanned automatically via river - unless status != 'active'
